+++ bumplus/__init__.py
@@ -1,12 +1,10 @@
+"""Bumplus top-level module."""
 from __future__ import absolute_import, division, print_function
 
 from .bumplus import Bumplus
 from .cli import main
 from .errors import BumplusError, NotBumplusDir, VersionNotDefined
 from .version import __version__
-
-__metaclass__ = type
-
 
 # Why do we export them again?  Because flake8 will report F401 if `__all__` is
 # absent.  We have to declare `__all__` explicitly.  See [1] for details.