+++ bumplus/constants.py
@@ -1,5 +1,4 @@
+"""Constants."""
 from __future__ import absolute_import, division, print_function
-
-__metaclass__ = type
 
 CONF_FILE = ".bumplus.toml"