+++ bumplus/filters.py
@@ -1,10 +1,17 @@
+"""Template filters."""
 from __future__ import absolute_import, division, print_function
 
-__metaclass__ = type
 
+def strftime(value, format_spec="%x"):
+    """Template filter to format data and time.
 
-def strftime(value, format="%x"):
-    return value.strftime(format)
+    :param value:
+        A datatime object.
+    :param format_spec:
+        A format string which is the same as the parameter of ``strftime``
+        method of Python ``datatime`` object.
+    """
+    return value.strftime(format_spec)
 
 
 filters = {"strftime": strftime}